Разделение кода на части
