+++ js/data.js
@@ -1,0 +1,8 @@
+const offer = {
+  types: ['palace', 'flat', 'house', 'bungalow'],
+  times: ['12:00', '13:00', '14:00'],
+  features: ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'],
+  photos: ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'],
+}
+
+export { offer };
